Test cases replaced and description and readme updated
